P4-OVS: Fix issue in Makefile.am
When P4-OVS was using a older version of GRPC, we are manually applying
PATCH-01-GRPC for few distributions. When GRPC version is updated to latest
we dont need this patch anymore and this file has been removed. But the
same is not removed from Makefile.am, which causes build errors on a
fresh pull.

Change-type: DefectResolution
Title: P4-OVS: Fix issue in Makefile.am
Signed-off-by: n-sandeep <sandeep.nagapattinam@intel.com>