P4-OVS: Fix issues with members assigned to a group
When a two digit member is pass to a group, ovs-p4ctl is splitting
this two digit number into two single digit numbers.
  Eg: member ID 23 is splited into member ID 2 and member ID 3

This is causing problem in programming backend and behavior seems unexpected.
This review is raised to fix this problem in splitting member ID's passed to
a group.

Change-type: DefectResolution
Title: Fix issues with members assigned to a group
HSDES-number: 14016303413
HSDES-tenant: Server_Platf_LAN.Bug
Signed-off-by: n-sandeep <sandeep.nagapattinam@intel.com>