Update default action to drop
For a packet, in a partifcular table if no rule has been matched,
by default drop the packet.
Existing code sends this packet to DEFAULT_EXCEPTION_VPORT or
DEFAULT_MGMT_VPORT, which reaches VM0 (vhost-user-0) for LNW.
This causes arp to be resolved even when control plane rules
are deleted.

With this fix, we drop the packet when no rules are matched.

Signed-off-by: n-sandeep <sandeep.nagapattinam@intel.com>